+++ test/support/mocks.ex
@@ -1,1 +1,3 @@
-Mox.defmock(MockClusterHandler, for: TdCluster.ClusterHandlerBehaviour)
+Mox.defmock(ElasticsearchMock, for: Elasticsearch.API)
+Mox.defmock(MockClusterHandler, for: TdCluster.ClusterHandler)
+Mox.defmock(MockIndexWorker, for: TdCore.Search.IndexWorker)